data-pipeline/requirements.txt
pandas==2.1.4
numpy==1.26.3
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.3
pytest==7.4.4
pytest-cov==4.1.0
